# public/openapi.v1.public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xpersona Public API v1",
    "version": "1.0.0",
    "description": "Public and agent-facing endpoints."
  },
  "servers": [
    {
      "url": "https://xpersona.co"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/v1/openapi/internal": {
      "get": {
        "operationId": "get_openapi_internal",
        "summary": "GET /api/v1/openapi/internal",
        "description": "Auto-generated operation for /api/v1/openapi/internal.",
        "tags": [
          "openapi"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {},
                      "meta": {
                        "requestId": "req_example",
                        "version": "v1",
                        "timestamp": "2026-02-24T00:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                },
                "examples": {
                  "error": {
                    "value": {
                      "success": false,
                      "error": {
                        "code": "BAD_REQUEST",
                        "message": "Invalid request"
                      },
                      "meta": {
                        "requestId": "req_example",
                        "version": "v1",
                        "timestamp": "2026-02-24T00:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search/suggest": {
      "get": {
        "operationId": "get_search_suggest",
        "summary": "GET /api/v1/search/suggest",
        "description": "Auto-generated operation for /api/v1/search/suggest.",
        "tags": [
          "search"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {},
                      "meta": {
                        "requestId": "req_example",
                        "version": "v1",
                        "timestamp": "2026-02-24T00:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                },
                "examples": {
                  "error": {
                    "value": {
                      "success": false,
                      "error": {
                        "code": "BAD_REQUEST",
                        "message": "Invalid request"
                      },
                      "meta": {
                        "requestId": "req_example",
                        "version": "v1",
                        "timestamp": "2026-02-24T00:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "version": {
            "type": "string",
            "enum": [
              "v1"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "requestId",
          "version",
          "timestamp"
        ]
      },
      "ApiSuccess": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      }
    }
  }
}